main/app/services/jwt/token.py
from datetime import timedelta, datetime
from jose import jwt
from config import Configs


def create_access_token(data: dict, expires_delta: timedelta | None = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=Configs.ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, Configs.SECRET, algorithm=Configs.ALGORITHM)
    return encoded_jwt


